refactor(web): SSR 지원을 위한 탐색 쿼리 헬퍼 추가
